fixed filtering issue
