+++ it.baeyens.arduino.core/src/it/baeyens/arduino/ui/LibraryPreferencePage.java
@@ -1,7 +1,9 @@
 package it.baeyens.arduino.ui;
 
-import java.util.HashMap;
-import java.util.Map;
+import java.util.Collection;
+import java.util.Set;
+import java.util.TreeMap;
+import java.util.TreeSet;
 
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.IStatus;
@@ -26,7 +28,6 @@ import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Control;
 import org.eclipse.swt.widgets.Text;
-import org.eclipse.swt.widgets.Tree;
 import org.eclipse.swt.widgets.TreeColumn;
 import org.eclipse.swt.widgets.TreeItem;
 import org.eclipse.ui.IWorkbench;
@@ -35,15 +36,16 @@ import org.eclipse.ui.dialogs.FilteredTree;
 import org.eclipse.ui.dialogs.PatternFilter;
 
 import it.baeyens.arduino.managers.Library;
-import it.baeyens.arduino.managers.LibraryTree;
+import it.baeyens.arduino.managers.LibraryIndex;
 import it.baeyens.arduino.managers.Manager;
 
 public class LibraryPreferencePage extends PreferencePage implements IWorkbenchPreferencePage {
 
-	private Tree tree;
-	private CheckboxTreeViewer viewer;
+	private FilteredTree tree;
+	private TreeViewer viewer;
 	private TreeEditor editor;
-	
+	private LibraryTree libs = new LibraryTree();
+
 	@Override
 	public void init(IWorkbench workbench) {
 		// nothing needed here
@@ -51,7 +53,8 @@ public class LibraryPreferencePage extends PreferencePage implements IWorkbenchP
 
 	@Override
 	protected void performDefaults() {
-		viewer.setInput(new LibraryTree());
+		// viewer.setInput(new LibraryTree());
+		libs.reset();
 		editor.getEditor().dispose();
 		super.performDefaults();
 	}
@@ -73,28 +76,21 @@ public class LibraryPreferencePage extends PreferencePage implements IWorkbenchP
 
 	@Override
 	public boolean performOk() {
-		final HashMap<String, String> libs = new HashMap<>();
-		for (TreeItem category : tree.getItems()) {
-			for (TreeItem  item : category.getItems()) {
-				if (item.getData() instanceof LibraryTree.Library) {
-					libs.put(item.getText(0), item.getText(1));
-				}
-			}
-		}
 		new Job(Messages.ui_Adopting_arduino_libraries) {
 
 			@Override
 			protected IStatus run(IProgressMonitor monitor) {
-				MultiStatus status = new MultiStatus(Activator.getId(), 0, Messages.ui_installing_arduino_libraries, null);
-				for (Map.Entry<String, String> lib : libs.entrySet()) {
-						Library toRemove = Manager.getLibraryIndex().getInstalledLibrary(lib.getKey());
-						if (toRemove != null && !toRemove.getVersion().equals(lib.getValue())) {
-							status.add(toRemove.remove(monitor));
-						}
-						Library toInstall = Manager.getLibraryIndex().getLibrary(lib.getKey(), lib.getValue());
-						if (toInstall != null && !toInstall.isInstalled()) {
-							status.add(toInstall.install(monitor));
-						}
+				MultiStatus status = new MultiStatus(Activator.getId(), 0, Messages.ui_installing_arduino_libraries,
+						null);
+				for (LibraryTree.Library lib : libs.getAllLibraries()) {
+					Library toRemove = Manager.getLibraryIndex().getInstalledLibrary(lib.getName());
+					if (toRemove != null && !toRemove.getVersion().equals(lib.getVersion())) {
+						status.add(toRemove.remove(monitor));
+					}
+					Library toInstall = Manager.getLibraryIndex().getLibrary(lib.getName(), lib.getVersion());
+					if (toInstall != null && !toInstall.isInstalled()) {
+						status.add(toInstall.install(monitor));
+					}
 				}
 				return status;
 			}
@@ -102,14 +98,14 @@ public class LibraryPreferencePage extends PreferencePage implements IWorkbenchP
 
 		return true;
 	}
-	
+
 	public void createTree(Composite parent) {
 		// filtering applied to all columns
 		PatternFilter filter = new PatternFilter() {
 			protected boolean isLeafMatch(final Viewer viewer, final Object element) {
-				TreeViewer treeViewer = (TreeViewer) viewer;
-				int numberOfColumns = treeViewer.getTree().getColumnCount();
-				ITableLabelProvider labelProvider = (ITableLabelProvider) treeViewer.getLabelProvider();
+
+				int numberOfColumns = ((TreeViewer) viewer).getTree().getColumnCount();
+				ITableLabelProvider labelProvider = (ITableLabelProvider) ((TreeViewer) viewer).getLabelProvider();
 				boolean isMatch = false;
 				for (int columnIndex = 0; columnIndex < numberOfColumns; columnIndex++) {
 					String labelText = labelProvider.getColumnText(element, columnIndex);
@@ -118,15 +114,19 @@ public class LibraryPreferencePage extends PreferencePage implements IWorkbenchP
 				return isMatch;
 			}
 		};
-		
-		tree = new FilteredTree(parent, SWT.CHECK | SWT.V_SCROLL | SWT.BORDER | SWT.FULL_SELECTION, filter, true).getViewer().getTree();
 
-		// trick to replace the tree viewer
-		viewer = new CheckboxTreeViewer(tree);
-		viewer.setLabelProvider(new LibraryLabelProvider());
-		viewer.setContentProvider(new LibraryContentProvider());
-		viewer.setCheckStateProvider(new LibraryCheckProvider());
-		viewer.setInput(new LibraryTree());
+		tree = new FilteredTree(parent, SWT.CHECK | SWT.V_SCROLL | SWT.BORDER | SWT.FULL_SELECTION, filter, true) {
+			protected TreeViewer doCreateTreeViewer(Composite composite, int style) {
+				CheckboxTreeViewer viewer = new CheckboxTreeViewer(composite);
+				viewer.setCheckStateProvider(new LibraryCheckProvider());
+				viewer.setLabelProvider(new LibraryLabelProvider());
+				viewer.setContentProvider(new LibraryContentProvider());
+				return viewer;
+			}
+
+		};
+		viewer = tree.getViewer();
+		viewer.setInput(libs);
 
 		TreeColumn name = new TreeColumn(viewer.getTree(), SWT.LEFT);
 		name.setWidth(300);
@@ -151,11 +151,18 @@ public class LibraryPreferencePage extends PreferencePage implements IWorkbenchP
 				}
 				final TreeItem item = event.item instanceof TreeItem ? (TreeItem) event.item : null;
 				if (item != null && event.detail == SWT.CHECK) {
-					if (item.getItemCount() > 0) {
+					if (item.getData() instanceof LibraryTree.Category) {
 						item.setGrayed(false);
+						for (LibraryTree.Library child : ((LibraryTree.Category) item.getData()).getLibraries()) {
+							if (item.getChecked()) {
+								child.setVersion(child.getLatest());
+							} else {
+								child.setVersion(null);
+							}
+						}
 						for (TreeItem child : item.getItems()) {
 							child.setChecked(item.getChecked());
-							if (child.getChecked()) {
+							if (item.getChecked()) {
 								child.setText(1, ((LibraryTree.Library) child.getData()).getLatest());
 							} else {
 								child.setText(1, "");
@@ -163,8 +170,11 @@ public class LibraryPreferencePage extends PreferencePage implements IWorkbenchP
 						}
 					} else {
 						if (item.getChecked()) {
+							((LibraryTree.Library) item.getData())
+									.setVersion(((LibraryTree.Library) item.getData()).getLatest());
 							item.setText(1, ((LibraryTree.Library) item.getData()).getLatest());
 						} else {
+							((LibraryTree.Library) item.getData()).setVersion(null);
 							item.setText(1, "");
 						}
 						verifySubtreeCheckStatus(item.getParentItem());
@@ -191,6 +201,7 @@ public class LibraryPreferencePage extends PreferencePage implements IWorkbenchP
 					// cell
 					combo.addSelectionListener(new SelectionAdapter() {
 						public void widgetSelected(SelectionEvent event) {
+							((LibraryTree.Library) item.getData()).setVersion(combo.getText());
 							item.setText(1, combo.getText());
 							// Item selected: end the editing session
 							combo.dispose();
@@ -202,10 +213,11 @@ public class LibraryPreferencePage extends PreferencePage implements IWorkbenchP
 	}
 
 	/**
-	 * Ensures the correct checked/unchecked/greyed attributes are set
-	 * on the category.
+	 * Ensures the correct checked/unchecked/greyed attributes are set on the
+	 * category.
 	 * 
-	 * @param item the tree item representing the category
+	 * @param item
+	 *            the tree item representing the category
 	 */
 	private void verifySubtreeCheckStatus(TreeItem item) {
 		boolean grayed = false;
@@ -257,7 +269,7 @@ public class LibraryPreferencePage extends PreferencePage implements IWorkbenchP
 				return ((LibraryTree.Node) element).getName();
 			case 1:
 				if (element instanceof LibraryTree.Library) {
-					return ((LibraryTree.Library) element).getInstalled();
+					return ((LibraryTree.Library) element).getVersion();
 				} else {
 					return "";
 				}
@@ -274,10 +286,10 @@ public class LibraryPreferencePage extends PreferencePage implements IWorkbenchP
 		@Override
 		public boolean isChecked(Object element) {
 			if (element instanceof LibraryTree.Library) {
-				return ((LibraryTree.Library) element).getInstalled() != null;
+				return ((LibraryTree.Library) element).getVersion() != null;
 			} else if (element instanceof LibraryTree.Category) {
 				for (LibraryTree.Library library : ((LibraryTree.Category) element).getLibraries()) {
-					if (library.getInstalled() != null) {
+					if (library.getVersion() != null) {
 						return true;
 					}
 				}
@@ -289,7 +301,7 @@ public class LibraryPreferencePage extends PreferencePage implements IWorkbenchP
 		public boolean isGrayed(Object element) {
 			if (element instanceof LibraryTree.Category && isChecked(element)) {
 				for (LibraryTree.Library library : ((LibraryTree.Category) element).getLibraries()) {
-					if (library.getInstalled() == null) {
+					if (library.getVersion() == null) {
 						return true;
 					}
 				}
@@ -339,6 +351,189 @@ public class LibraryPreferencePage extends PreferencePage implements IWorkbenchP
 
 		@Override
 		public void inputChanged(Viewer arg0, Object arg1, Object arg2) {
+		}
+	}
+
+	private static class LibraryTree {
+		private TreeMap<String, Category> categories = new TreeMap<>();
+
+		public static interface Node {
+			boolean hasChildren();
+
+			Object[] getChildren();
+
+			Object getParent();
+
+			String getName();
+		}
+
+		public class Category implements Comparable<Category>, Node {
+			private String name;
+			private TreeMap<String, Library> libraries = new TreeMap<>();
+
+			public Category(String name) {
+				this.name = name;
+			}
+
+			public String getName() {
+				return name;
+			}
+
+			public Collection<Library> getLibraries() {
+				return libraries.values();
+			}
+
+			@Override
+			public int compareTo(Category other) {
+				return name.compareTo(other.name);
+			}
+
+			@Override
+			public boolean hasChildren() {
+				return !libraries.isEmpty();
+			}
+
+			@Override
+			public Object[] getChildren() {
+				return libraries.values().toArray();
+			}
+
+			@Override
+			public Object getParent() {
+				return LibraryTree.this;
+			}
+		}
+
+		public class Library implements Comparable<Library>, Node {
+			private String name;
+			private Category category;
+			private TreeSet<Version> versions = new TreeSet<>();
+			private String version;
+
+			public Library(Category category, String name) {
+				this.category = category;
+				this.name = name;
+			}
+
+			public Collection<Version> getVersions() {
+				return versions;
+			}
+
+			public String getName() {
+				return name;
+			}
+
+			public String getLatest() {
+				return versions.last().toString();
+			}
+
+			public String getVersion() {
+				return version;
+			}
+
+			public void setVersion(String version) {
+				this.version = version;
+			}
+
+			@Override
+			public int compareTo(Library other) {
+				return name.compareTo(other.name);
+			}
+
+			@Override
+			public boolean hasChildren() {
+				return false;
+			}
+
+			@Override
+			public Object[] getChildren() {
+				return null;
+			}
+
+			@Override
+			public Object getParent() {
+				return category;
+			}
+		}
+
+		public class Version implements Comparable<Object> {
+			private String[] parts;
+
+			public Version(String version) {
+				parts = version.split("\\.");
+			}
+
+			@Override
+			public int compareTo(Object other) {
+				if (other instanceof String) {
+					return this.compareTo(new Version((String) other));
+				} else if (other instanceof Version) {
+					return this.compareParts(((Version) other).parts, 0);
+				} else {
+					throw new UnsupportedOperationException();
+				}
+			}
+
+			private int compareParts(String[] other, int level) {
+				if (parts.length > level && other.length > level) {
+					if (parts[level].compareTo(other[level]) == 0) {
+						return this.compareParts(other, ++level);
+					} else {
+						try {
+							return new Integer(parts[level]).compareTo(Integer.parseInt(other[level]));
+						} catch (Exception e) {
+							return parts[level].compareTo(other[level]);
+						}
+					}
+				} else {
+					return parts.length > other.length ? 1 : -1;
+				}
+			}
+
+			public String toString() {
+				return String.join(".", parts);
+			}
+		}
+
+		public LibraryTree() {
+			LibraryIndex libraryIndex = Manager.getLibraryIndex();
+
+			for (String categoryName : libraryIndex.getCategories()) {
+				Category category = new Category(categoryName);
+				for (it.baeyens.arduino.managers.Library library : libraryIndex.getLibraries(categoryName)) {
+					Library lib = category.libraries.get(library.getName());
+					if (lib == null) {
+						lib = new Library(category, library.getName());
+						category.libraries.put(lib.getName(), lib);
+					}
+					lib.versions.add(new Version(library.getVersion()));
+					if (library.isInstalled()) {
+						lib.version = library.getVersion();
+					}
+				}
+
+				categories.put(category.getName(), category);
+			}
+		}
+
+		public Collection<Category> getCategories() {
+			return categories.values();
+		}
+
+		public Collection<Library> getAllLibraries() {
+			Set<Library> all = new TreeSet<>();
+			for (Category category : categories.values()) {
+				all.addAll(category.getLibraries());
+			}
+			return all;
+		}
+
+		public void reset() {
+			LibraryIndex libraryIndex = Manager.getLibraryIndex();
+			for (Library library : this.getAllLibraries()) {
+				it.baeyens.arduino.managers.Library installed = libraryIndex.getInstalledLibrary(library.getName());
+				library.setVersion(installed != null ? installed.getVersion() : null);
+			}
 		}
 	}
 }
